rename console printer
