+++ src/console-printer.js
